requirements.txt
build==1.0.3
click==8.1.7
packaging==23.2
pyproject_hooks==1.0.0
python-dotenv==1.0.0
